Ajax Loader and Green Up Sections.
Started working on Green Up project tracker view, grabbed remote links
from repositories on my machine, created small helper js function to
parse the dom (though I might just pull in jQuery to make things
easier/more readable... can't remember if Harp supports coffee script or
not.)

Next step is to pull in each of the latest commits for the 4
repositories and also to provide links to them. This will be done via js
and via using ul rel="" with classes of latest-commit so that the
javascript can be generic.